+++ src/grscheller/datastructures/__init__.py
@@ -14,7 +14,7 @@
 
 """Classes to implement algorithms, functional and imperative styles supported."""
 
-__version__ = "0.13.0"
+__version__ = "0.13.1"
 __author__ = "Geoffrey R. Scheller"
 __copyright__ = "Copyright (c) 2023-2024 Geoffrey R. Scheller"
 __license__ = "Apache License 2.0"